hanyang 6th beginner > F

+++ Backjoon/17946/src.cpp
@@ -6,8 +6,8 @@ int main(void){
 	int C;
 	scanf("%d", &C);
 	while(C--){
-		long long int res = 0, n;
-		scanf("%lld", &n);
-		printf("%lld\n",1);
+		int res = 0, n;
+		scanf("%d", &n);
+		printf("%d\n",1);
 	}
 }
